SIMPLY-2536 Fix Vexing Parse warning
The warning Xcode 11.3.1 gave suggested a possible bug: `parentheses were disambiguated as redundant parentheses around declaration of variable named '_mutex' [-Wvexing-parse]`.
However, the project uses the `_` as a naming convention for class members, and this class does define a `_mutex` member. So it seems that the intention was to use that member to initialize a local mutex, as done in other files of this project.
Clang++ provides 3 suggestions:
```
note: add a variable name to declare a 'std::unique_lock<std::mutex>' initialized with '_mutex'
note: add enclosing parentheses to perform a function-style cast
note: remove parentheses to silence this warning
```
While the 3rd is just warning suppression and the 2nd seems nonsensical in this context, the 1st achieves the alleged intention.

Also see: https://en.wikipedia.org/wiki/Most_vexing_parse