SIMPLY-2536 Fix redundant-move warning
Warning from Xcode 11.3.1: `redundant move in return statement [-Wredundant-move]`

Since we are returning a param, its lvalue is definitely lost, but at the same time the compiler needs to treat it as an rvalue in order to return it. Therefore no matter what it will perform an implicit move, making our explicit move redundant.
Good explanation here: https://developers.redhat.com/blog/2019/04/12/understanding-when-not-to-stdmove-in-c/